Move integration test and add backend check (#5)
The integration test no longer builds and uses the binary
but runs the Cobra command directly. Also, an additional
check is added that makes sure the backend Terraform was
initialized with matches the backend configuration of the
current run. This avoids that state is accidentally pushed
to the wrong backend when switching environments.