Remove non-functional cache dir change for tests
Setting XDG_CACHE_HOME to a temporary directory for tests did not work
because the directory is initialized in an init function. Setting the
environment variable comes to late. We'd have to explicitly call
xdg.Reload() to fix this. However, we can just remove this. On CI, the
Terraform binary is never cached anyways, and locally we want it to be
cached in the default directory so tests run fast. Having to download
Terraform on every test execution would be painful.